+++ config_default.php
@@ -393,3 +393,10 @@ $config['cops_home_page'] = 'INDEX';
  * Note: author, language and publisher are always assumed present and cannot be filtered here
  */
 $config['cops_show_not_set_filter'] = ['custom', 'rating', 'series', 'tag'];
+
+/*
+ * Show links to filter by Author, Language, Publisher, Rating, Serie or Tag on page detail
+ * 1 : Yes (enable)
+ * 0 : No
+ */
+$config['cops_show_filter_links'] = '1';