+++ lib/Pages/PageTagDetail.php
@@ -24,6 +24,7 @@ class PageTagDetail extends Page
         }
         $this->idPage = $tag->getEntryId();
         $this->title = $tag->getTitle();
+        $this->currentUri = $tag->getUri();
         $this->parentTitle = localize("tags.title");
         $this->parentUri = $tag->getParentUri();
     }